+++ src/felupe/field/_base.py
@@ -98,9 +98,7 @@ class Field:
     """
 
     def __init__(self, region, dim=1, values=0.0, dtype=None, **kwargs):
-        self.region = region
-        self.dim = dim
-        self.shape = self.region.quadrature.npoints, self.region.mesh.ncells
+        self.__field__ = Field
 
         # set optional user-defined attributes
         for key, value in kwargs.items():
@@ -124,10 +122,49 @@ class Field:
                 shape=(region.mesh.npoints, dim), fill_value=values, dtype=dtype
             )
 
-        eai, ai = self._indices_per_cell(self.region.mesh.cells, dim)
-        self.indices = Indices(eai, ai, region, dim)
+        self.dim = dim
+        self.reload(region=region)
 
-        self.__field__ = Field
+    def reload(self, region, mode="edge", **kwargs):
+        """Reload the field with a new region in-place. The field values are extended
+        to the new region if necessary.
+
+        Parameters
+        ----------
+        region : Region
+            The new region for the field.
+        mode : str, optional
+            The padding mode to use if the field values need to be extended. Default is
+            "edge".
+        **kwargs : dict, optional
+            Additional keyword arguments are passed to :func:`numpy.pad`.
+        """
+
+        if region is not None:
+            self.region = region
+            self.shape = self.region.quadrature.npoints, self.region.mesh.ncells
+
+            shape = (self.region.mesh.npoints, self.dim)
+            if shape != self.values.shape:
+                length = shape[0] - self.values.shape[0]
+
+                if length >= 0:
+                    pad_width = ((0, length), (0, 0))
+
+                    # expand values
+                    self.values = np.pad(
+                        self.values,
+                        pad_width=pad_width[: len(self.values.shape)],
+                        mode=mode,
+                        **kwargs,
+                    )
+
+                else:  # trim values
+                    self.values = self.values[: length]
+
+
+            eai, ai = self._indices_per_cell(self.region.mesh.cells, self.dim)
+            self.indices = Indices(eai, ai, region, self.dim)
 
     def _indices_per_cell(self, cells, dim):
         "Calculate pre-defined indices for sparse matrices."